empleados/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-ckeditor==6.0.0
django-js-asset==1.2.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
Unipath==1.1
